requirements.txt
aiohttp>=3.7.0,<3.8.0
python-dotenv>=0.19.0,<2.0.0